scraper/function/requirements.txt
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
cryptography==36.0.1
idna==3.3
pdfminer.six==20211012
pycparser==2.21
requests==2.27.1
urllib3==1.26.8